+++ test/multi_domain_secure_server_test.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:convert';
 import 'dart:io';
 import 'dart:typed_data';
 
@@ -47,6 +48,36 @@ void main() {
       expect(socketError, isA<HandshakeException>());
     });
 
+    test('isValidHostname', () {
+      // Valid intranet hostnames:
+      expect(MultiDomainSecureServer.isValidHostname('a'), isTrue);
+      expect(MultiDomainSecureServer.isValidHostname('localhost'), isTrue);
+      expect(MultiDomainSecureServer.isValidHostname('foo-bar'), isTrue);
+      expect(MultiDomainSecureServer.isValidHostname('my-server-123'), isTrue);
+
+      // Valid general domains
+      expect(MultiDomainSecureServer.isValidHostname('a.x'), isTrue);
+      expect(MultiDomainSecureServer.isValidHostname('a.com'), isTrue);
+      expect(MultiDomainSecureServer.isValidHostname('example.com'), isTrue);
+      expect(MultiDomainSecureServer.isValidHostname('a.example.com'), isTrue);
+      expect(
+          MultiDomainSecureServer.isValidHostname('sub.example.org'), isTrue);
+
+      // Invalid cases
+      expect(MultiDomainSecureServer.isValidHostname('-localhost'), isFalse);
+      expect(MultiDomainSecureServer.isValidHostname('localhost-'), isFalse);
+      expect(MultiDomainSecureServer.isValidHostname('-localhost-'), isFalse);
+      expect(MultiDomainSecureServer.isValidHostname('example..com'), isFalse);
+      expect(MultiDomainSecureServer.isValidHostname('.example.com'), isFalse);
+      expect(MultiDomainSecureServer.isValidHostname('example.com.'), isFalse);
+      expect(MultiDomainSecureServer.isValidHostname('exa mple.com'), isFalse);
+      expect(MultiDomainSecureServer.isValidHostname('example@com'), isFalse);
+      expect(
+          MultiDomainSecureServer.isValidHostname(
+              'toolongtoolongtooooooolongtooloooooooongtoolongtoolongtoooloooooongtooloooooong.com'),
+          isFalse);
+    });
+
     test('parseSNIHostname', () {
       var clientHello1 = Uint8List.fromList(
           '22 3 1 1 60 1 0 1 56 3 3 61 102 193 32 209 174 53 83 188 33 75 119 88 137 2 44 45 206 110 171 116 31 225 143 242 253 243 48 100 205 70 97 32 44 215 108 123 47 114 71 97 202 224 139 71 86 115 204 215 150 142 34 2 242 186 145 113 134 164 74 82 139 166 48 41 0 98 19 2 19 3 19 1 192 48 192 44 192 40 192 36 192 20 192 10 0 159 0 107 0 57 204 169 204 168 204 170 255 133 0 196 0 136 0 129 0 157 0 61 0 53 0 192 0 132 192 47 192 43 192 39 192 35 192 19 192 9 0 158 0 103 0 51 0 190 0 69 0 156 0 60 0 47 0 186 0 65 192 17 192 7 0 5 0 4 192 18 192 8 0 22 0 10 0 255 1 0 0 141 0 43 0 9 8 3 4 3 3 3 2 3 1 0 51 0 38 0 36 0 29 0 32 184 1 34 43 208 205 96 129 192 251 165 124 120 253 218 236 32 6 190 57 24 207 103 41 146 78 97 177 153 195 174 2 0 0 0 16 0 14 0 0 11 102 111 111 111 98 97 114 46 99 111 109 0 11 0 2 1 0 0 10 0 10 0 8 0 29 0 23 0 24 0 25 0 13 0 24 0 22 8 6 6 1 6 3 8 5 5 1 5 3 8 4 4 1 4 3 2 1 2 3 0 16 0 14 0 12 2 104 50 8 104 116 116 112 47 49 46 49'
@@ -74,6 +105,12 @@ void main() {
 
       var hostname3 = MultiDomainSecureServer.parseSNIHostname(clientHello3);
       expect(hostname3, equals('foooooooo.com'));
+
+      var clientHello4 = base64.decode(
+          'FgMBAgABAAH8AwMCcVJ2MrnTm5G+khJM/h0U5VjUiS7UMqMJIuGcqCyt2CDZslFp9zMiXQVjM114d43Q5fQaxbG2/t5R9eoBWpL2mwAsSkoTARMCEwPALMArzKnAMMAvzKjACsAJwBTAEwCdAJwANQAvwAjAEgAKVgABAAGHCgoAAAAAABAADgAAC2Zvb29iYXIuY29tABcAAP8BAAEAAAoADAAKmpoAHQAXABgAGQALAAIBAAAQAAsACQhodHRwLzEuMQAFAAUBAAAAAAANABgAFgQDCAQEAQUDAgMIBQgFBQEIBgYBAgEAEgAAADMAKwApmpoAAQAAHQAgbex07GLXM5A2tHVM/firBbYS9Fo1HpeJs1ueeZNw3TEALQACAQEAKwALCtraAwQDAwMCAwEAGwADAgABamoAAQAAFQDEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA==');
+
+      var hostname4 = MultiDomainSecureServer.parseSNIHostname(clientHello4);
+      expect(hostname4, equals('fooobar.com'));
     });
   });
 }